+++ CHANGELOG.md
@@ -15,7 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 
 ### Fixed 🐞
 
--   Labels do not take delta height into account([#1523](https://github.com/MaibornWolff/codecharta/issues/1523))
+-   Labels do not take delta height into account ([#1523](https://github.com/MaibornWolff/codecharta/issues/1523))
 
 ### Chore 👨‍💻 👩‍💻